+++ service_crategen/src/commands/generate/codegen/error_types.rs
@@ -53,22 +53,6 @@ pub trait GenerateErrorTypes {
                 }}
 
                 {error_from_body_impl}
-                {error_from_type_impl}
-                impl From<CredentialsError> for {type_name} {{
-                    fn from(err: CredentialsError) -> {type_name} {{
-                        {type_name}::Credentials(err)
-                    }}
-                }}
-                impl From<HttpDispatchError> for {type_name} {{
-                    fn from(err: HttpDispatchError) -> {type_name} {{
-                        {type_name}::HttpDispatch(err)
-                    }}
-                }}
-                impl From<io::Error> for {type_name} {{
-                    fn from(err: io::Error) -> {type_name} {{
-                        {type_name}::HttpDispatch(HttpDispatchError::from(err))
-                    }}
-                }}
                 impl fmt::Display for {type_name} {{
                     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {{
                         write!(f, \"{{}}\", self.description())
@@ -85,7 +69,6 @@ pub trait GenerateErrorTypes {
             type_name = error_type_name(service, operation_name),
             error_from_body_impl =
                 self.generate_error_from_body_impl(operation_name, operation, service),
-            error_from_type_impl = self.generate_error_from_type_impl(operation_name, service),
             error_types = self
                 .generate_error_enum_types(operation, error_documentation)
                 .unwrap_or_else(|| String::from("")),
@@ -106,7 +89,7 @@ pub trait GenerateErrorTypes {
         if operation.errors.is_some() {
             for error in operation.errors() {
                 // some botocore definitions include Validation in every errors list, some take it as assumed
-                // skip it if it's listed, as we implement it for all error types below
+                // skip it if it's listed, as we implement it for all error types in the RusotoError enum.
                 if error.idiomatic_error_name() != "Validation" {
                     enum_types.push(format!(
                         "\n{}\n{}(String)",
@@ -121,22 +104,6 @@ pub trait GenerateErrorTypes {
             }
         }
 
-        enum_types.push(
-            "/// An error occurred dispatching the HTTP request\nHttpDispatch(HttpDispatchError)"
-                .to_string(),
-        );
-        enum_types.push(
-            "/// An error was encountered with AWS credentials.\nCredentials(CredentialsError)"
-                .to_string(),
-        );
-        enum_types.push(
-            "/// A validation error occurred.  Details from AWS are provided.\nValidation(String)"
-                .to_string(),
-        );
-        enum_types.push(
-            "/// An error occurred parsing the response payload.\nParseError(String)".to_string(),
-        );
-        enum_types.push("/// An unknown error occurred.  The raw HTTP response is provided.\nUnknown(BufferedHttpResponse)".to_string());
         Some(enum_types.join(","))
     }
 
@@ -148,7 +115,6 @@ pub trait GenerateErrorTypes {
         service: &Service,
     ) -> Option<String> {
         let mut type_matchers: Vec<String> = Vec::new();
-        let error_type = error_type_name(service, operation_name);
 
         if operation.errors.is_some() {
             // botocore has some dulicated errors
@@ -165,26 +131,6 @@ pub trait GenerateErrorTypes {
             }
         }
 
-        type_matchers.push(format!(
-            "{error_type}::Validation(ref cause) => cause",
-            error_type = error_type
-        ));
-        type_matchers.push(format!(
-            "{error_type}::Credentials(ref err) => err.description()",
-            error_type = error_type
-        ));
-        type_matchers.push(format!(
-            "{error_type}::HttpDispatch(ref dispatch_error) => dispatch_error.description()",
-            error_type = error_type
-        ));
-        type_matchers.push(format!(
-            "{error_type}::ParseError(ref cause) => cause",
-            error_type = error_type
-        ));
-        type_matchers.push(format!(
-            "{error_type}::Unknown(_) => \"unknown error\"",
-            error_type = error_type
-        ));
         Some(type_matchers.join(",\n"))
     }
 
@@ -194,7 +140,6 @@ pub trait GenerateErrorTypes {
         operation: &Operation,
         service: &Service,
     ) -> String;
-    fn generate_error_from_type_impl(&self, operation_name: &str, service: &Service) -> String;
 }
 
 pub struct RestJsonErrorTypes;
@@ -210,7 +155,7 @@ impl GenerateErrorTypes for XmlErrorTypes {
     ) -> String {
         format!("
                 impl {type_name} {{
-                    pub fn from_response(res: BufferedHttpResponse) -> {type_name} {{
+                    pub fn from_response(res: BufferedHttpResponse) -> RusotoError<{type_name}> {{
                         {{
                             let reader = EventReader::new(res.body.as_slice());
                             let mut stack = XmlResponse::new(reader.into_iter().peekable());
@@ -221,7 +166,7 @@ impl GenerateErrorTypes for XmlErrorTypes {
                                 }}
                            }}
                        }}
-                       {type_name}::Unknown(res)
+                       RusotoError::Unknown(res)
                     }}
 
                     fn deserialize<T>(stack: &mut T) -> Result<XmlError, XmlParseError> where T: Peek + Next {{
@@ -231,19 +176,6 @@ impl GenerateErrorTypes for XmlErrorTypes {
                 type_name = error_type_name(service, operation_name),
                 error_deserializer = self.generate_error_deserializer(service),
                 type_matchers = self.generate_error_type_matchers(operation_name, operation, service))
-    }
-
-    fn generate_error_from_type_impl(&self, operation_name: &str, service: &Service) -> String {
-        format!(
-            "
-                impl From<XmlParseError> for {type_name} {{
-                    fn from(err: XmlParseError) -> {type_name} {{
-                        let XmlParseError(message) = err;
-                        {type_name}::ParseError(message.to_string())
-                    }}
-                }}",
-            type_name = error_type_name(service, operation_name)
-        )
     }
 }
 
@@ -288,7 +220,7 @@ impl XmlErrorTypes {
                     .as_ref()
                     .and_then(|http_error| http_error.code.as_ref())
                     .unwrap_or(&error.shape);
-                type_matchers.push(format!("\"{error_code}\" => return {error_type}::{error_name}(String::from(parsed_error.message))",
+                type_matchers.push(format!("\"{error_code}\" => return RusotoError::Service({error_type}::{error_name}(String::from(parsed_error.message)))",
                     error_code = error_code,
                     error_type = error_type,
                     error_name = error.idiomatic_error_name()))
@@ -309,7 +241,7 @@ impl GenerateErrorTypes for JsonErrorTypes {
     ) -> String {
         format!("
                 impl {type_name} {{
-                    pub fn from_response(res: BufferedHttpResponse) -> {type_name} {{
+                    pub fn from_response(res: BufferedHttpResponse) -> RusotoError<{type_name}> {{
                         if let Ok(json) = from_slice::<SerdeJsonValue>(&res.body) {{
                             let raw_error_type = json.get(\"__type\").and_then(|e| e.as_str()).unwrap_or(\"Unknown\");
                             let error_message = json.get(\"message\").and_then(|m| m.as_str()).unwrap_or(\"\");
@@ -321,23 +253,11 @@ impl GenerateErrorTypes for JsonErrorTypes {
                                 {type_matchers}
                             }}
                         }}
-                        return {type_name}::Unknown(res);
+                        return RusotoError::Unknown(res);
                     }}
                 }}",
                 type_name = error_type_name(service, operation_name),
                 type_matchers = self.generate_error_type_matchers(operation_name, operation, service))
-    }
-
-    fn generate_error_from_type_impl(&self, operation_name: &str, service: &Service) -> String {
-        format!(
-            "
-                impl From<serde_json::error::Error> for {type_name} {{
-                    fn from(err: serde_json::error::Error) -> {type_name} {{
-                        {type_name}::ParseError(err.description().to_string())
-                    }}
-                }}",
-            type_name = error_type_name(service, operation_name)
-        )
     }
 }
 
@@ -356,7 +276,7 @@ impl JsonErrorTypes {
         if operation.errors.is_some() {
             for error in operation.errors() {
                 if error.shape != "ValidationException" {
-                    type_matchers.push(format!("\"{error_shape}\" => return {error_type}::{error_name}(String::from(error_message))",
+                    type_matchers.push(format!("\"{error_shape}\" => return RusotoError::Service({error_type}::{error_name}(String::from(error_message)))",
                         error_shape = error.shape,
                         error_type = error_type,
                         error_name = error.idiomatic_error_name()))
@@ -364,8 +284,7 @@ impl JsonErrorTypes {
             }
         }
         type_matchers.push(format!(
-            "\"ValidationException\" => return {error_type}::Validation(error_message.to_string())",
-            error_type = error_type
+            "\"ValidationException\" => return RusotoError::Validation(error_message.to_string())",
         ));
         type_matchers.push(format!("_ => {{}}"));
         type_matchers.join(",\n")
@@ -383,7 +302,7 @@ impl GenerateErrorTypes for RestJsonErrorTypes {
                 impl {type_name} {{
                     // see boto RestJSONParser impl for parsing errors
                     // https://github.com/boto/botocore/blob/4dff78c840403d1d17db9b3f800b20d3bd9fbf9f/botocore/parsers.py#L838-L850
-                    pub fn from_response(res: BufferedHttpResponse) -> {type_name} {{
+                    pub fn from_response(res: BufferedHttpResponse) -> RusotoError<{type_name}> {{
                         if let Ok(json) = from_slice::<SerdeJsonValue>(&res.body) {{
                             let error_type = match res.headers.get(\"x-amzn-errortype\") {{
                                 Some(raw_error_type) => {{
@@ -401,14 +320,10 @@ impl GenerateErrorTypes for RestJsonErrorTypes {
                                 {type_matchers}
                             }}
                         }}
-                        return {type_name}::Unknown(res);
+                        return RusotoError::Unknown(res);
                     }}
                 }}",
                 type_name = error_type_name(service, operation_name),
                 type_matchers = JsonErrorTypes.generate_error_type_matchers(operation_name, operation, service))
-    }
-
-    fn generate_error_from_type_impl(&self, operation_name: &str, service: &Service) -> String {
-        JsonErrorTypes.generate_error_from_type_impl(operation_name, service)
     }
 }